[Osquery] Fix small issues (#141083)
(cherry picked from commit ccbac37155d6544a9f579e47052deb6a4eb69786)
