+++ x-pack/plugins/osquery/public/lens/view_results_in_lens.tsx
@@ -1,0 +1,234 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import React, { useCallback, useMemo } from 'react';
+import { i18n } from '@kbn/i18n';
+import { EuiButtonEmpty, EuiButtonIcon, EuiToolTip } from '@elastic/eui';
+import type {
+  PersistedIndexPatternLayer,
+  PieVisualizationState,
+  TermsIndexPatternColumn,
+  TypedLensByValueInput,
+} from '@kbn/lens-plugin/public';
+import { DOCUMENT_FIELD_NAME as RECORDS_FIELD } from '@kbn/lens-plugin/common/constants';
+import { FilterStateStore } from '@kbn/es-query';
+import { ViewResultsActionButtonType } from '../live_queries/form/pack_queries_status_table';
+import type { LogsDataView } from '../common/hooks/use_logs_data_view';
+import { useKibana } from '../common/lib/kibana';
+import { useLogsDataView } from '../common/hooks/use_logs_data_view';
+
+interface ViewResultsInLensActionProps {
+  actionId?: string;
+  agentIds?: string[];
+  buttonType: ViewResultsActionButtonType;
+  endDate?: string;
+  startDate?: string;
+  mode?: string;
+}
+
+const ViewResultsInLensActionComponent: React.FC<ViewResultsInLensActionProps> = ({
+  actionId,
+  agentIds,
+  buttonType,
+  endDate,
+  startDate,
+  mode,
+}) => {
+  const lensService = useKibana().services.lens;
+  const { data: logsDataView } = useLogsDataView({ skip: !actionId });
+
+  const handleClick = useCallback(
+    (event) => {
+      event.preventDefault();
+
+      if (logsDataView) {
+        lensService?.navigateToPrefilledEditor(
+          {
+            id: '',
+            timeRange: {
+              from: startDate ?? 'now-1d',
+              to: endDate ?? 'now',
+              mode: mode ?? (startDate || endDate) ? 'absolute' : 'relative',
+            },
+            attributes: getLensAttributes(logsDataView, actionId, agentIds),
+          },
+          {
+            openInNewTab: true,
+            skipAppLeave: true,
+          }
+        );
+      }
+    },
+    [actionId, agentIds, endDate, lensService, logsDataView, mode, startDate]
+  );
+
+  const isDisabled = useMemo(() => !actionId || !logsDataView, [actionId, logsDataView]);
+
+  if (buttonType === ViewResultsActionButtonType.button) {
+    return (
+      <EuiButtonEmpty size="xs" iconType="lensApp" onClick={handleClick} isDisabled={isDisabled}>
+        {VIEW_IN_LENS}
+      </EuiButtonEmpty>
+    );
+  }
+
+  return (
+    <EuiToolTip content={VIEW_IN_LENS}>
+      <EuiButtonIcon
+        iconType="lensApp"
+        disabled={false}
+        onClick={handleClick}
+        aria-label={VIEW_IN_LENS}
+        isDisabled={isDisabled}
+      />
+    </EuiToolTip>
+  );
+};
+
+function getLensAttributes(
+  logsDataView: LogsDataView,
+  actionId?: string,
+  agentIds?: string[]
+): TypedLensByValueInput['attributes'] {
+  const dataLayer: PersistedIndexPatternLayer = {
+    columnOrder: ['8690befd-fd69-4246-af4a-dd485d2a3b38', 'ed999e9d-204c-465b-897f-fe1a125b39ed'],
+    columns: {
+      '8690befd-fd69-4246-af4a-dd485d2a3b38': {
+        sourceField: 'type',
+        isBucketed: true,
+        dataType: 'string',
+        scale: 'ordinal',
+        operationType: 'terms',
+        label: 'Top values of type',
+        params: {
+          otherBucket: true,
+          size: 5,
+          missingBucket: false,
+          orderBy: {
+            columnId: 'ed999e9d-204c-465b-897f-fe1a125b39ed',
+            type: 'column',
+          },
+          orderDirection: 'desc',
+        },
+      } as TermsIndexPatternColumn,
+      'ed999e9d-204c-465b-897f-fe1a125b39ed': {
+        sourceField: RECORDS_FIELD,
+        isBucketed: false,
+        dataType: 'number',
+        scale: 'ratio',
+        operationType: 'count',
+        label: 'Count of records',
+      },
+    },
+    incompleteColumns: {},
+  };
+
+  const xyConfig: PieVisualizationState = {
+    shape: 'pie',
+    layers: [
+      {
+        layerType: 'data',
+        legendDisplay: 'default',
+        nestedLegend: false,
+        layerId: 'layer1',
+        metric: 'ed999e9d-204c-465b-897f-fe1a125b39ed',
+        numberDisplay: 'percent',
+        primaryGroups: ['8690befd-fd69-4246-af4a-dd485d2a3b38'],
+        categoryDisplay: 'default',
+      },
+    ],
+  };
+
+  const agentIdsQuery = agentIds?.length
+    ? {
+        bool: {
+          minimum_should_match: 1,
+          should: agentIds?.map((agentId) => ({ match_phrase: { 'agent.id': agentId } })),
+        },
+      }
+    : undefined;
+
+  return {
+    visualizationType: 'lnsPie',
+    title: `Action ${actionId} results`,
+    references: [
+      {
+        id: logsDataView.id,
+        name: 'indexpattern-datasource-current-indexpattern',
+        type: 'index-pattern',
+      },
+      {
+        id: logsDataView.id,
+        name: 'indexpattern-datasource-layer-layer1',
+        type: 'index-pattern',
+      },
+      {
+        name: 'filter-index-pattern-0',
+        id: logsDataView.id,
+        type: 'index-pattern',
+      },
+    ],
+    state: {
+      datasourceStates: {
+        indexpattern: {
+          layers: {
+            layer1: dataLayer,
+          },
+        },
+      },
+      filters: [
+        {
+          $state: { store: FilterStateStore.APP_STATE },
+          meta: {
+            index: 'filter-index-pattern-0',
+            negate: false,
+            alias: null,
+            disabled: false,
+            params: {
+              query: actionId,
+            },
+            type: 'phrase',
+            key: 'action_id',
+          },
+          query: {
+            match_phrase: {
+              action_id: actionId,
+            },
+          },
+        },
+        ...(agentIdsQuery
+          ? [
+              {
+                $state: { store: FilterStateStore.APP_STATE },
+                meta: {
+                  alias: 'agent IDs',
+                  disabled: false,
+                  index: 'filter-index-pattern-0',
+                  key: 'query',
+                  negate: false,
+                  type: 'custom',
+                  value: JSON.stringify(agentIdsQuery),
+                },
+                query: agentIdsQuery,
+              },
+            ]
+          : []),
+      ],
+      query: { language: 'kuery', query: '' },
+      visualization: xyConfig,
+    },
+  };
+}
+
+const VIEW_IN_LENS = i18n.translate(
+  'xpack.osquery.pack.queriesTable.viewLensResultsActionAriaLabel',
+  {
+    defaultMessage: 'View in Lens',
+  }
+);
+
+export const ViewResultsInLensAction = React.memo(ViewResultsInLensActionComponent);
