+++ x-pack/plugins/osquery/public/live_queries/form/pack_results_header.tsx
@@ -35,6 +35,7 @@ const StyledIconsList = styled(EuiFlexItem)`
 
 export const PackResultsHeader = ({ actionId, addToCase }: PackResultsHeadersProps) => (
   <>
+    <EuiSpacer size={'l'} />
     <EuiFlexGroup direction="row" gutterSize="m">
       <StyledResultsHeading grow={false}>
         <EuiText>
