+++ x-pack/plugins/osquery/public/shared_components/osquery_action/index.tsx
@@ -6,17 +6,12 @@
  */
 
 import { EuiLoadingContent, EuiEmptyPrompt, EuiCode } from '@elastic/eui';
-import React, { useMemo } from 'react';
+import React from 'react';
 
 import { FormattedMessage } from '@kbn/i18n-react';
+import { OsqueryEmptyPrompt, OsqueryNotAvailablePrompt } from '../prompts';
 import type { AddToTimelinePayload } from '../../timelines/get_add_to_timeline';
-import {
-  AGENT_STATUS_ERROR,
-  EMPTY_PROMPT,
-  NOT_AVAILABLE,
-  PERMISSION_DENIED,
-  SHORT_EMPTY_TITLE,
-} from './translations';
+import { AGENT_STATUS_ERROR, PERMISSION_DENIED, SHORT_EMPTY_TITLE } from './translations';
 import { useKibana } from '../../common/lib/kibana';
 import { LiveQuery } from '../../live_queries';
 import { OsqueryIcon } from '../../components/osquery_icon';
@@ -39,22 +34,11 @@ const OsqueryActionComponent: React.FC<OsqueryActionProps> = ({
 }) => {
   const permissions = useKibana().services.application.capabilities.osquery;
 
-  const emptyPrompt = useMemo(
-    () => (
-      <EuiEmptyPrompt
-        icon={<OsqueryIcon />}
-        title={<h2>{SHORT_EMPTY_TITLE}</h2>}
-        titleSize="xs"
-        body={<p>{EMPTY_PROMPT}</p>}
-      />
-    ),
-    []
-  );
   const { osqueryAvailable, agentFetched, isLoading, policyFetched, policyLoading, agentData } =
     useIsOsqueryAvailable(agentId);
 
   if (agentId && agentFetched && !agentData) {
-    return emptyPrompt;
+    return <OsqueryEmptyPrompt />;
   }
 
   if (
@@ -91,14 +75,7 @@ const OsqueryActionComponent: React.FC<OsqueryActionProps> = ({
   }
 
   if (agentId && !osqueryAvailable) {
-    return (
-      <EuiEmptyPrompt
-        icon={<OsqueryIcon />}
-        title={<h2>{SHORT_EMPTY_TITLE}</h2>}
-        titleSize="xs"
-        body={<p>{NOT_AVAILABLE}</p>}
-      />
-    );
+    return <OsqueryNotAvailablePrompt />;
   }
 
   if (agentId && agentData?.status !== 'online') {
