+++ x-pack/plugins/osquery/public/shared_components/osquery_results/osquery_result.test.tsx
@@ -20,7 +20,7 @@ import {
   DETAILS_ID,
   DETAILS_QUERY,
   DETAILS_TIMESTAMP,
-  mockCasesContext,
+  getMockedKibanaConfig,
 } from './test_utils';
 
 jest.mock('../../common/lib/kibana');
@@ -43,34 +43,8 @@ const defaultProps = {
   queryId: '',
 };
 const mockKibana = (permissionType: unknown = defaultPermissions) => {
-  useKibanaMock.mockReturnValue({
-    services: {
-      application: {
-        capabilities: permissionType,
-      },
-      cases: {
-        helpers: {
-          canUseCases: jest.fn(),
-        },
-        ui: {
-          getCasesContext: jest.fn().mockImplementation(() => mockCasesContext),
-        },
-      },
-      data: {
-        dataViews: {
-          getCanSaveSync: jest.fn(),
-          hasData: {
-            hasESData: jest.fn(),
-            hasUserDataView: jest.fn(),
-            hasDataView: jest.fn(),
-          },
-        },
-      },
-      notifications: {
-        toasts: jest.fn(),
-      },
-    },
-  } as unknown as ReturnType<typeof useKibana>);
+  const mockedKibana = getMockedKibanaConfig(permissionType);
+  useKibanaMock.mockReturnValue(mockedKibana);
 };
 
 const renderWithContext = (Element: React.ReactElement) =>
