+++ x-pack/plugins/osquery/public/shared_components/osquery_results/osquery_results.test.tsx
@@ -17,7 +17,7 @@ import * as useAllLiveQueries from '../../actions/use_all_live_queries';
 import * as useLiveQueryDetails from '../../actions/use_live_query_details';
 import { PERMISSION_DENIED } from '../osquery_action/translations';
 import * as privileges from '../../action_results/use_action_privileges';
-import { defaultLiveQueryDetails, DETAILS_QUERY, mockCasesContext } from './test_utils';
+import { defaultLiveQueryDetails, DETAILS_QUERY, getMockedKibanaConfig } from './test_utils';
 
 jest.mock('../../common/lib/kibana');
 
@@ -49,34 +49,8 @@ const defaultPermissions = {
 };
 
 const mockKibana = (permissionType: unknown = defaultPermissions) => {
-  useKibanaMock.mockReturnValue({
-    services: {
-      application: {
-        capabilities: permissionType,
-      },
-      cases: {
-        helpers: {
-          canUseCases: jest.fn(),
-        },
-        ui: {
-          getCasesContext: jest.fn().mockImplementation(() => mockCasesContext),
-        },
-      },
-      data: {
-        dataViews: {
-          getCanSaveSync: jest.fn(),
-          hasData: {
-            hasESData: jest.fn(),
-            hasUserDataView: jest.fn(),
-            hasDataView: jest.fn(),
-          },
-        },
-      },
-      notifications: {
-        toasts: jest.fn(),
-      },
-    },
-  } as unknown as ReturnType<typeof useKibana>);
+  const mockedKibana = getMockedKibanaConfig(permissionType);
+  useKibanaMock.mockReturnValue(mockedKibana);
 };
 
 const renderWithContext = (Element: React.ReactElement) =>
