--- x-pack/plugins/security_solution/public/common/components/markdown_editor/plugins/osquery/utils.ts
@@ -1,31 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { isEmpty, reduce } from 'lodash';
-
-export const convertECSMappingToObject = (
-  ecsMapping: Array<{
-    key: string;
-    result: {
-      type: string;
-      value: string;
-    };
-  }>
-) =>
-  reduce(
-    ecsMapping,
-    (acc, value) => {
-      if (!isEmpty(value?.key) && !isEmpty(value.result?.type) && !isEmpty(value.result?.value)) {
-        acc[value.key] = {
-          [value.result.type]: value.result.value,
-        };
-      }
-
-      return acc;
-    },
-    {} as Record<string, { field?: string; value?: string }>
-  );
