+++ x-pack/test/osquery_cypress/artifact_manager.ts
@@ -6,5 +6,5 @@
  */
 
 export async function getLatestVersion(): Promise<string> {
-  return '8.5.0-SNAPSHOT';
+  return '8.6.0-SNAPSHOT';
 }
